extract icon mapping in dedicated mapper
