+++ src/screens/retaliation/RetaliationCardViewModelMapper.ts
@@ -1,15 +1,12 @@
-import { ImageSourcePropType } from 'react-native'
-import {
-  Retaliation,
-  RetaliationSiteType,
-} from '../../core/entities/Retaliation'
+import { RetaliationSocialIconImageMapper } from './../home/retaliation/RetaliationSocialIconImageMapper'
+import { Retaliation } from '../../core/entities/Retaliation'
 import { RetaliationCardViewModel } from './RetaliationCardViewModel'
 
 export const RetaliationCardViewModelMapper = {
   map: (retaliation: Retaliation): RetaliationCardViewModel => {
     return {
       id: retaliation.id,
-      socialIcon: mapRetaliationIconImage(
+      socialIcon: RetaliationSocialIconImageMapper.map(
         retaliation.openGraph?.site ?? 'others',
       ),
       image: retaliation.openGraph?.image ?? '',
@@ -18,17 +15,4 @@ export const RetaliationCardViewModelMapper = {
       url: retaliation.sourceUrl,
     }
   },
-}
-
-function mapRetaliationIconImage(
-  site: RetaliationSiteType,
-): ImageSourcePropType {
-  switch (site) {
-    case 'facebook':
-      return require('../../assets/images/facebook.png')
-    case 'twitter':
-      return require('../../assets/images/twitter.png')
-    case 'others':
-      return require('../../assets/images/otherSocialNetwork.png')
-  }
 }
